Direct upload of CAFM prime MS Word file
The aim of this primer is to provide sufficient operational information for a user to start working with advanced conductive atomic force microscopy (CAFM) procedures, such as tomography and constant bias time-domain measurements. It is most relevant to Bruker ICON/Veeco Dimension instruments, although many of the practises will be generally applicable to AFM/CAFM measurements. It is worth noting that the documentation/help files for Nanoscope and Nanoscope Analysis are very detailed and extremely useful. I would thoroughly recommend making use of them. I have not referenced material within the text, but I have included a selection of useful references at the end that are relevant to much of the technical discussion.

The details contained in this document combine the training I received when I started my EngD at UCL in 2013 with my experience using AFM and CAFM. Much of the information is based on approaches that I found worked during my experimental research, but they were not necessarily based on well-defined methodologies (although they were generally based on some modicum of science and/or scientific practise). As such, if you find alternative, more effective approaches, then I would recommend using them!

This primer describes measurement processes, from module and probe setup through software implementation and data analysis. It generally concerns contact-mode measurements (typical for CAFM) although some tapping-mode procedures will be mentioned. Notably, this is not an exhaustive training document for CAFM. Rather, it is written for users who already have some experience with AFM at least, in that they do not need to learn the theory and have a reasonable grasp of a typical scanning probe workflow. It should be used as a reference document rather than a training manual. Ideally, the user should take the information contain within and develop their own practical understanding and expertise, for which there is no substitute or alternative click-and-go approach. However, there should be sufficient detail here for a user to follow the workflow they need for their measurement.

Please email me with any questions or concerns, as I’ve certainly forgotten some details, described things poorly/incorrectly, or left some parts otherwise incomplete.

Enjoy!